Move exported classes out of internal packages
