+++ dev/io.openliberty.microprofile.metrics.5.0.internal/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2022 IBM Corporation and others.
+# Copyright (c) 2022, 2023 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License 2.0
 # which accompanies this distribution, and is available at 	
@@ -35,10 +35,12 @@ Import-Package: \
 
 Private-Package: \
 	com.ibm.ws.microprofile.metrics.resources,\
+	io.openliberty.microprofile.metrics50.internal,\
 	com.ibm.ws.microprofile.metrics.exceptions
 
 Export-Package: \
-	io.openliberty.microprofile.metrics50.internal,\
+	io.openliberty.microprofile.metrics50,\
+	io.openliberty.microprofile.metrics50.rest,\
 	io.openliberty.microprofile.metrics50.helper,\
 	io.openliberty.smallrye.metrics.adapters,\
 	io.openliberty.smallrye.metrics.cdi.adapters
@@ -46,7 +48,7 @@ Export-Package: \
 -dsannotations-inherit: true
 -dsannotations: \
 	io.openliberty.microprofile.metrics50.internal.MetricsConfig,\
-	io.openliberty.microprofile.metrics50.internal.SharedMetricRegistries,\
+	io.openliberty.microprofile.metrics50.SharedMetricRegistries,\
 	io.openliberty.microprofile.metrics50.internal.SmallryeMetricsCDIMetadata,\
 	io.openliberty.microprofile.metrics50.internal.ApplicationListener50
