+++ dev/io.openliberty.microprofile.metrics.5.0.internal/src/io/openliberty/microprofile/metrics50/rest/MetricRESTHandler.java
@@ -1,16 +1,16 @@
 /*******************************************************************************
- * Copyright (c) 2020, 2022 IBM Corporation and others.
+ * Copyright (c) 2020, 2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package io.openliberty.microprofile.metrics50.internal;
+package io.openliberty.microprofile.metrics50.rest;
 
 import java.io.IOException;
 import java.lang.reflect.Constructor;
@@ -29,6 +29,8 @@ import com.ibm.wsspi.rest.handler.RESTHandler;
 import com.ibm.wsspi.rest.handler.RESTRequest;
 import com.ibm.wsspi.rest.handler.RESTResponse;
 
+import io.openliberty.microprofile.metrics50.SharedMetricRegistries;
+import io.openliberty.microprofile.metrics50.helper.Constants;
 import io.openliberty.microprofile.metrics50.helper.Util;
 import jakarta.servlet.http.HttpServletResponse;
