+++ dev/io.openliberty.microprofile.metrics.5.0.public.internal/src/io/openliberty/microprofile/metrics/internal/publicapi/PublicMetricsRESTProxyServlet.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018, 2022 IBM Corporation and others.
+ * Copyright (c) 2018, 2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -12,7 +12,7 @@
  *******************************************************************************/
 package io.openliberty.microprofile.metrics.internal.publicapi;
 
-import io.openliberty.microprofile.metrics50.internal.BaseMetricsRESTProxyServlet;
+import io.openliberty.microprofile.metrics50.rest.BaseMetricsRESTProxyServlet;
 
 public class PublicMetricsRESTProxyServlet extends BaseMetricsRESTProxyServlet {
     private static final long serialVersionUID = 1L;
