+++ dev/io.openliberty.restfulWS.mpMetrics.filter/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2022 IBM Corporation and others.
+# Copyright (c) 2022, 2023 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License 2.0
 # which accompanies this distribution, and is available at
@@ -47,7 +47,7 @@ Import-Package: \
     jakarta.servlet; version="[6.0,7)", \
     org.eclipse.microprofile.metrics;version="[5.0,6)",\
 	org.eclipse.microprofile.metrics.*;version="[5.0,6)",\
-	io.openliberty.microprofile.metrics50.internal,\
+	io.openliberty.microprofile.metrics50,\
     !*.internal.*,*
 
 Private-Package: \
